+++ src/slave(Pi-Pico)/src/global_vars.h
@@ -2,14 +2,15 @@
 #define GLOBAL_VARS_H
 
 #include <stdbool.h> //* to use boolean data type
+#include <stdint.h>  //* to use uint8_t, uint16_t, uint32, etc. data types
 
-extern const unsigned short int GREEN_LED;
+extern const uint8_t GREEN_LED;
 extern const bool GPIO_ON;
 extern const bool GPIO_OFF;
 
-extern const short int MOTORs_PINs_motor_right_terminal_1;
-extern const short int MOTORs_PINs_motor_right_terminal_2;
-extern const short int MOTORs_PINs_motor_left_terminal_1;
-extern const short int MOTORs_PINs_motor_left_terminal_2;
+extern const uint8_t MOTORs_PINs_motor_right_terminal_1;
+extern const uint8_t MOTORs_PINs_motor_right_terminal_2;
+extern const uint8_t MOTORs_PINs_motor_left_terminal_1;
+extern const uint8_t MOTORs_PINs_motor_left_terminal_2;
 
 #endif
